Added WelcomeUserState + fix Exception message + improved Linq query
WelcomeUser project:
Implemented same architecture of the EchoBotWithCounter project. The project now contains a WelcomeUserState class handled by WelcomeUserStateAccessors.

line 62 all Startup.cs
botConfig in the InvalidOperationException message is always null, it would be better for the user to provide the path in which the .bot file shold be.

Line 66 all Startup.cs
Improved LINQ query