Minor correction to docs
After running through this tutorial, I noticed a couple things that 
weren't quite right.  Firstly, the directory name for MsbotClone was
cased incorrectly, meaning it didn't work when trying to run on my
Mac.  Secondly, the name of the .bot file didn't match what was
stored in the sample folder.